Restructure config module (#109)
* Restructure config module

* Update CHANGELOG.md

* Fix new 1.78 clippy errors
